Added branch out after login
The design and logic will be changed later, but currently for easier designing there's a dropdown button with the option for catering or beauty industry. There's a login simulation where, for now, we pretend that the log in happened. After pressing the log in with google button there's redirection to empty catering and beauty main pages.